requirements.txt
google-genai>=1.5.0
flask>=2.0.0
Pillow>=9.0.0
werkzeug>=2.0.0
requests>=2.25.0
gunicorn>=20.1.0